// lib/types/quotes.ts
export type AssetMeta = {
  symbol: string;
  price: number;
  previousClose: number;
  change: number;
  changePercent: number;
  currency: string;
  marketState: string;
  dayHigh: number;
  dayLow: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  volume: number;
  averageDailyVolume3Month: number;
  marketCap: number;
  trailingPE: number;
  forwardPE: number;
  priceToBook: number;
  trailingEps: number;
  forwardEps: number;
  trailingAnnualDividendRate: number;
  trailingAnnualDividendYield: number;
  dividendDate: number | null;
  exDividendDate: number | null;
  shortName: string;
  longName: string;
  exchange: string;
  quoteType: string;
  beta: number;
};